# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/tenant/delete": {
      "post": {
        "operationId": "TenantDeletionController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDeletionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Request deletion of the calling tenant (MSA § 7). 30-day grace.",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "TenantDeletionController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read the calling tenant's current deletion request, if any",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/tenant/delete/{id}": {
      "delete": {
        "operationId": "TenantDeletionController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel a pending deletion request (only requested/scheduled state)",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/billing/checkout-session": {
      "post": {
        "operationId": "BillingAdminController_checkoutSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a Stripe Checkout session for self-serve tier purchase",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/billing/portal-session": {
      "post": {
        "operationId": "BillingAdminController_portalSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a Stripe Customer Portal session for the calling tenant",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/billing/seats": {
      "post": {
        "operationId": "BillingAdminController_addSeats",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Self-serve seat increase (writes through Stripe with prorations)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Billing Rules Platform API",
    "description": "Pre-flight a claim against payer × state × code rules.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id"
      }
    },
    "schemas": {
...
      "RequestDeletionDto": {
        "type": "object",
        "properties": {}
      },
...
      "CheckoutSessionDto": {
        "type": "object",
        "properties": {}
      },
      "PortalSessionDto": {
        "type": "object",
        "properties": {}
      },
      "SeatChangeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}